build: remove eachdist.py
eachdist.py duplicated functionality tox already provides
(install/test/lint/format per package), which is the core complaint
in #1462. Its remaining real responsibilities are split into two
small, single-purpose scripts instead of one multi-command tool:

- scripts/repo_targets.py: stdlib-only distribution discovery, used
  by scripts/griffe_check.py for the public API breaking-change check.
- scripts/release.py: version bumping used by the release workflows
  and .github/scripts/update-version*.sh, reading the renamed
  repo.ini (formerly eachdist.ini).

Also removes the coverage tox env and scripts/coverage.sh: it isn't
in tox's envlist, isn't run by any workflow, and already referenced
instrumentation/* and exporter-datadog paths that no longer exist in
this repo since the split to opentelemetry-python-contrib.
