+++ .github/scripts/update-version.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 
-sed -i "/\[stable\]/{n;s/version=.*/version=$1/}" eachdist.ini
-sed -i "/\[prerelease\]/{n;s/version=.*/version=$2/}" eachdist.ini
+sed -i "/\[stable\]/{n;s/version=.*/version=$1/}" repo.ini
+sed -i "/\[prerelease\]/{n;s/version=.*/version=$2/}" repo.ini
 
-./scripts/eachdist.py update_versions --versions stable,prerelease
+./scripts/release.py update_versions --versions stable,prerelease
