+++ repo.ini
@@ -43,11 +43,3 @@ packages=
     opentelemetry-semantic-conventions
     opentelemetry-test-utils
     tests
-
-[lintroots]
-extraroots=examples/*,scripts/
-subglob=*.py,tests/,test/,src/*,examples/*
-
-[testroots]
-extraroots=examples/*,tests/
-subglob=tests/,test/
