+++ tox.ini
@@ -128,8 +128,6 @@ envlist =
 [testenv]
 deps =
   lint: -r dev-requirements.txt
-  coverage: pytest
-  coverage: pytest-cov
 
   api: -r {toxinidir}/opentelemetry-api/test-requirements.txt
 
@@ -216,10 +214,6 @@ setenv =
   CONTRIB_REPO_INSTRUMENTATION_WSGI={env:CONTRIB_REPO_INSTRUMENTATION_WSGI:{env:CONTRIB_REPO}\#egg=opentelemetry-instrumentation-wsgi&subdirectory=instrumentation/opentelemetry-instrumentation-wsgi}
   CONTRIB_REPO_INSTRUMENTATION_FLASK={env:CONTRIB_REPO_INSTRUMENTATION_FLASK:{env:CONTRIB_REPO}\#egg=opentelemetry-instrumentation-flask&subdirectory=instrumentation/opentelemetry-instrumentation-flask}
   UV_CONFIG_FILE={toxinidir}/tox-uv.toml
-commands_pre =
-  ; In order to get a healthy coverage report,
-  ; we have to install packages in editable mode.
-  coverage: python {toxinidir}/scripts/eachdist.py install --editable
 
 commands =
   test-opentelemetry-api: pytest {toxinidir}/opentelemetry-api/tests {posargs}
@@ -300,8 +294,6 @@ commands =
 
   test-opentelemetry-test-utils: pytest {toxinidir}/tests/opentelemetry-test-utils/tests {posargs}
   lint-opentelemetry-test-utils: sh -c "cd tests && pylint --rcfile ../.pylintrc {toxinidir}/tests/opentelemetry-test-utils"
-
-  coverage: {toxinidir}/scripts/coverage.sh
 
 [testenv:spellcheck]
 recreate = True
@@ -393,7 +385,6 @@ recreate = True
 deps =
   GitPython==3.1.50
   griffe==1.7.3
-  toml
 commands =
   ; griffe check before to fail fast if there are any issues
   python {toxinidir}/scripts/griffe_check.py
